+++ IntelPresentMon/AppCef/ipm-ui-vue/src/core/api.ts
@@ -54,10 +54,10 @@ export class Api {
     static async loadEnvVars(): Promise<EnvVars> {
         return await this.invokeEndpointFuture('loadEnvVars', {});
     }
-    static async introspect(): Promise<{metrics: Metric[], stats: Stat[], units: Unit[], systemDeviceId: number, defaultAdapterId: number}> {
+    static async introspect(): Promise<{metrics: Metric[], stats: Stat[], units: Unit[], adapters: Adapter[], systemDeviceId: number, defaultAdapterId: number}> {
         const introData = await this.invokeEndpointFuture('Introspect', {});
         if (!Array.isArray(introData.metrics) || !Array.isArray(introData.stats) ||
-            !Array.isArray(introData.units) || typeof introData.systemDeviceId !== 'number' ||
+            !Array.isArray(introData.units) || !Array.isArray(introData.adapters) || typeof introData.systemDeviceId !== 'number' ||
             typeof introData.defaultAdapterId !== 'number') {
             console.log("error intro call");
             throw new Error('Bad member type returned from introspect');
@@ -74,13 +74,6 @@ export class Api {
     static async getTopGpuProcess(blacklist: string[]): Promise<Process|null> {
         const {top} = await this.invokeEndpointFuture('getTopGpuProcess', {blacklist});
         return top;
-    }
-    static async enumerateAdapters(): Promise<Adapter[]> {
-        const {adapters} = await this.invokeEndpointFuture('EnumerateAdapters', {});
-        if (!Array.isArray(adapters)) {
-            throw new Error('Bad (non-array) type returned from enumerateAdapters');
-        }
-        return adapters;
     }
     static async bindHotkey(binding: Binding): Promise<void> {
         await this.invokeEndpointFuture('BindHotkey', binding);
