+++ IntelPresentMon/AppCef/source/util/KernelActionRegistration.cpp
@@ -6,7 +6,6 @@ namespace p2c::client::util::kact {
 		IpcActionRegistrator() { ::p2c::client::util::IpcInvocationManager::RegisterDispatchBinding<A>(); }
 	};
 
-	IpcActionRegistrator<kproc::kact::EnumerateAdapters> reg_ibind_EnumerateAdapters_;
 	IpcActionRegistrator<kproc::kact::Introspect> reg_ibind_Introspect_;
 	IpcActionRegistrator<kproc::kact::PushSpecification> reg_ibind_PushSpecification_;
 	IpcActionRegistrator<kproc::kact::SetCapture> reg_ibind_SetCapture_;
