+++ IntelPresentMon/ControlLib/TelemetryCoordinator.cpp
@@ -61,72 +61,6 @@ namespace pmon::tel
         BuildLogicalDevicesAndRoutes_();
     }
 
-    std::vector<TelemetryCoordinator::AdapterInfo> TelemetryCoordinator::EnumerateAdapters() const
-    {
-        std::vector<AdapterInfo> adapters;
-        adapters.reserve(logicalDevicesById_.size());
-
-        for (const auto& [logicalDeviceId, logicalDevice] : logicalDevicesById_) {
-            if (logicalDeviceId == ipc::kSystemDeviceId) {
-                continue;
-            }
-
-            AdapterInfo adapter{};
-            adapter.id = logicalDeviceId;
-            const auto fingerprint = ResolveLogicalDeviceFingerprint_(logicalDevice);
-            adapter.vendor = fingerprint.vendor;
-            if (!fingerprint.deviceName.empty()) {
-                adapter.name = fingerprint.deviceName;
-            }
-            else {
-                adapter.name = "UNKNOWN_GPU";
-            }
-
-            adapter.gpuMemorySize = QueryGpuMemorySize_(logicalDevice).value;
-            const auto requestQpc = util::GetCurrentTimestamp();
-            const auto populateStaticMetric = [&]<typename T>(PM_METRIC metricId, T& destination) {
-                if (metricId == PM_METRIC_GPU_MEM_SIZE) {
-                    if constexpr (std::is_same_v<T, uint64_t>) {
-                        destination = adapter.gpuMemorySize;
-                    }
-                    return;
-                }
-                if (!logicalDevice.routes.contains(metricId)) {
-                    return;
-                }
-
-                try {
-                    const auto value = PollMetricForRoute_(logicalDevice, metricId, 0, requestQpc);
-                    if (const auto pVal = std::get_if<T>(&value)) {
-                        destination = *pVal;
-                    }
-                    else {
-                        throw util::Except<>("Type mismatch while enumerating adapter static metric");
-                    }
-                }
-                catch (...) {
-                    pmlog_error(util::ReportException("Adapter static metric query failed"))
-                        .pmwatch(logicalDeviceId)
-                        .pmwatch((int)metricId);
-                }
-            };
-
-            populateStaticMetric(PM_METRIC_GPU_SUSTAINED_POWER_LIMIT, adapter.gpuSustainedPowerLimit);
-            populateStaticMetric(PM_METRIC_GPU_MEM_MAX_BANDWIDTH, adapter.gpuMemoryMaxBandwidth);
-
-            adapters.push_back(std::move(adapter));
-        }
-
-        std::sort(adapters.begin(), adapters.end(), [](const AdapterInfo& lhs, const AdapterInfo& rhs) {
-            if (lhs.gpuMemorySize != rhs.gpuMemorySize) {
-                return lhs.gpuMemorySize > rhs.gpuMemorySize;
-            }
-            return lhs.id < rhs.id;
-        });
-
-        return adapters;
-    }
-
     void TelemetryCoordinator::RegisterDevicesToIpc(ipc::ServiceComms& comms) const
     {
         bool cpuRegistered = false;
