+++ IntelPresentMon/PresentMonService/AllActions.h
@@ -1,13 +1,12 @@
-// GENERATED HEADER
-#pragma once
-#include "acts/EnumerateAdapters.h"
-#include "acts/FinishEtlLogging.h"
-#include "acts/OpenSession.h"
-#include "acts/ReportMetricUse.h"
-#include "acts/SetEtwFlushPeriod.h"
-#include "acts/SetTelemetryPeriod.h"
-#include "acts/StartEtlLogging.h"
-#include "acts/StartTracking.h"
-#include "acts/StopPlayback.h"
-#include "acts/StopTracking.h"
+// GENERATED HEADER 
+#pragma once 
+#include "acts/FinishEtlLogging.h" 
+#include "acts/OpenSession.h" 
+#include "acts/ReportMetricUse.h" 
+#include "acts/SetEtwFlushPeriod.h" 
+#include "acts/SetTelemetryPeriod.h" 
+#include "acts/StartEtlLogging.h" 
+#include "acts/StartTracking.h" 
+#include "acts/StopPlayback.h" 
+#include "acts/StopTracking.h" 
 #include "acts/ReportFrameReadProgress.h"
