+++ IntelPresentMon/PresentMonService/PMMainThread.cpp
@@ -130,7 +130,6 @@ void TelemetryThreadEntry_(Service* const srv, PresentMon* const pm, ipc::Servic
         pmon::tel::TelemetryCoordinator coordinator{ pm->GetGpuTelemetryPeriod() };
         coordinator.RegisterDevicesToIpc(*pComms);
         coordinator.PopulateStaticsToIpc(*pComms);
-        pm->SetTelemetryAdapters(coordinator.EnumerateAdapters());
         pmlog_info(std::format(
             "Finished populating telemetry introspection, {} seconds elapsed",
             timer.Mark()));
