+++ IntelPresentMon/PresentMonService/acts/StartTracking.h
@@ -1,5 +1,6 @@
 ﻿#pragma once
 #include "../../Interprocess/source/act/ActionHelper.h"
+#include "../ActionExecutionContext.h"
 #include "../../CommonUtilities/win/Utilities.h"
 #include <format>
 #include <unordered_set>
