+++ IntelPresentMon/PresentMonService/acts/StopTracking.h
@@ -1,5 +1,6 @@
 ﻿#pragma once
 #include "../../Interprocess/source/act/ActionHelper.h"
+#include "../ActionExecutionContext.h"
 #include <format>
 #include <unordered_set>
